test(usage): share local HTTP test server
